Completed Problem 07.10

+++ 07. Loops/05. Calculate!/Program.cs
@@ -7,7 +7,7 @@ class Program
         int n = int.Parse(Console.ReadLine());
         double x = double.Parse(Console.ReadLine());
         double sum = 1;
-        int factorial = 1;
+        long factorial = 1;
         double powX = 1; 
         for (int i = 1; i <= n; i++)
         {
